refactor: Refactor the code
